backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==21.2.0
idna==3.7
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
dj-database-url
